requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-ckeditor==6.7.1
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==3.0.0
django-js-asset==2.2.0
environs==11.0.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
marshmallow==3.21.1
multidict==6.0.5
packaging==24.0
pillow==10.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
requests==2.32.2
sqlparse==0.5.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
